sync sia-eks and sia-fargate tools
