--- provider/aws/sia-eks/cmd/siad/main_test.go
@@ -1,40 +1,0 @@
-//
-// Copyright The Athenz Authors
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-//
-
-package main
-
-import (
-	"os"
-	"testing"
-	"time"
-
-	"github.com/AthenZ/athenz/provider/aws/sia-eks/devel/metamock"
-)
-
-func setup() {
-
-	go metamock.StartMetaServer("127.0.0.1:5082")
-	time.Sleep(3 * time.Second)
-}
-
-func teardown() {}
-
-func TestMain(m *testing.M) {
-	setup()
-	code := m.Run()
-	teardown()
-	os.Exit(code)
-}
