--- provider/aws/sia-eks/cmd/ztsmock/main.go
@@ -1,32 +1,0 @@
-//
-// Copyright The Athenz Authors
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-//
-
-package main
-
-import (
-	"flag"
-
-	"github.com/AthenZ/athenz/provider/aws/sia-eks/devel/ztsmock"
-)
-
-func main() {
-	endpoint := flag.String("endpoint", "127.0.0.1:4080", "optional mock zts endpoint")
-	flag.Parse()
-
-	go ztsmock.StartZtsServer(*endpoint)
-
-	select {}
-}
