+++ provider/aws/sia-eks/devel/data/sia_no_service2
@@ -1,0 +1,14 @@
+{
+    "version": "1.0.0",
+    "service": "api",
+    "services": {
+        "ui": {}
+    },
+    "accounts": [
+        {
+            "domain": "athenz",
+            "user": "nobody",
+            "account": "123456789012"
+        }
+    ]
+}
