+++ provider/aws/sia-eks/devel/data/sia_no_services
@@ -1,0 +1,11 @@
+{
+    "version": "1.0.0",
+    "service": "api",
+    "accounts": [
+        {
+            "domain": "athenz",
+            "user": "nobody",
+            "account": "123456789012"
+        }
+    ]
+}
