+++ provider/aws/sia-eks/devel/data/sia_rotate_key
@@ -1,0 +1,12 @@
+{
+    "version": "1.0.0",
+    "service": "api",
+    "rotate_key": true,
+    "accounts": [
+        {
+            "domain": "athenz",
+            "user": "nobody",
+            "account": "123456789012"
+        }
+    ]
+}
